Updated Directory Structure So That All Calibration Scripts are in camera_calibration directory
