+++ scripts/camera_calibration/ICP/realsense_icp/camera_sim.py
