+++ scripts/camera_calibration/ICP/realsense_icp/hw_drag_with_apriltag.py
@@ -11,7 +11,7 @@ Description:
 
 # Setting path for imports
 import sys
-sys.path.append('../')
+sys.path.append('../../../')
 
 # Imports
 from pydrake.all import *
