+++ scripts/camera_calibration/ICP/realsense_icp/tracking_apriltag.py
@@ -8,7 +8,7 @@ Description:
 
 # Setting path for imports
 import sys
-sys.path.append('../')
+sys.path.append('../../../')
 
 # Imports
 import matplotlib.pyplot as plt
