+++ scripts/camera_calibration/ICP/vision_calibration_kinova/kinova_image.py
