+++ scripts/camera_calibration/ICP/vision_calibration_kinova/kinova_pointcloud.py
