+++ scripts/camera_calibration/ICP/vision_calibration_kinova/realsense_export_ply_example.py
