+++ scripts/camera_calibration/ICP/vision_calibration_kinova/test_workspace/kinova_vision_example.py
