+++ scripts/camera_calibration/ICP/vision_calibration_kinova/test_workspace/kortex_vision_intrinsics.py
