+++ scripts/camera_calibration/ICP/vision_calibration_kinova/test_workspace/main.py
