+++ scripts/camera_calibration/ICP/vision_calibration_kinova/test_workspace/open3d_point_cloud_generator.py
