+++ scripts/camera_calibration/ICP/vision_calibration_kinova/test_workspace/realsense_pointcloud_example.py
