+++ scripts/camera_calibration/ICP/vision_calibration_kinova/utilities.py
