+++ scripts/camera_calibration/ICP/vision_calibration_realsense/kinova_pointcloud.py
