+++ scripts/camera_calibration/ICP/vision_calibration_realsense/realsense_export_ply_example.py
