+++ scripts/camera_calibration/ICP/vision_calibration_realsense/realsense_image.py
