+++ scripts/camera_calibration/ICP/vision_calibration_realsense/test_workspace/kinova_vision_example.py
