+++ scripts/camera_calibration/ICP/vision_calibration_realsense/test_workspace/kortex_vision_intrinsics.py
