+++ scripts/camera_calibration/ICP/vision_calibration_realsense/test_workspace/open3d_point_cloud_generator.py
