+++ scripts/camera_calibration/ICP/vision_calibration_realsense/test_workspace/realsense_pointcloud_example.py
